requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.11
gunicorn==25.3.0
whitenoise==6.12.0
dj-database-url==3.1.2
python-dotenv==1.2.2
